refactor(auction): 카프카 컨슈머 통합
